\name{animalMutation}
\alias{animalMutation}
\docType{data}
\title{
Mutation Distances of Fitch and Margoliash (1967)
}
\description{
Minimum numbers of mutations required to interrelate pairs of cytochromes c.  
}
\usage{data("animalMutation")}
\details{
Mutation distances are determined from the amino acid sequences.  The rows and columns correspond in order to:
\describe{
\item{1.}{Man}
\item{2.}{Monkey (Macacus mulatta)}
\item{3.}{Dog}
\item{4.}{Horse}
\item{5.}{Donkey}
\item{6.}{Pig}
\item{7.}{Rabbit}
\item{8.}{Kangaroo (Canopus Canguru)}
\item{9.}{Pekin duck}
\item{10.}{Pigeon}
\item{11.}{Chicken}
\item{12.}{King penguin (Aptenodytes patagonica)}
\item{13.}{Snapping turtle (Chelydra serpentina)}
\item{14.}{Rattlesnake (Crotalus adamanteus)}
\item{15.}{Tuna}
\item{16.}{Screwworm fly (Haematobia irritans)}
\item{17.}{Moth (Samia cynthia)}
\item{18.}{Neurospora (crassa)}
\item{19.}{Saccaromyces (oviformis)}
\item{20.}{Candida (krusei)}
}
}
\references{
Fitch, Walter M., and Emanuel Margoliash. "Construction of phylogenetic trees." Science 155.3760 (1967): 279-284.}
\examples{
data(animalMutation)
}
\keyword{datasets}
